Update README.md (#11)
* Update README.md

Update for 64-bit support and build. Display is now semi-broken and the canvas gets shown in the right side bar with a wrong resolution, not in the full screen display.

* Update README.md

had the time to go through the readme in full, was a bit late last night when I created the pull request. You were absolutely right.

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

---------

Co-authored-by: itsmeow <itsmeow@itsmeow.dev>